refactor: Split action handler and main.rs for v1.5.0
Major code structure refactoring:

handler/action.rs (2,853 lines) -> handler/action/ module:
- mod.rs: Type definitions and dispatch
- navigation.rs: MoveUp, MoveDown, MoveToTop, MoveToBottom
- tree_ops.rs: Expand, Collapse, ToggleExpand, CollapseAll, ExpandAll
- selection.rs: ToggleMark, ClearMarks, Copy, Cut
- file_ops.rs: Paste, ConfirmDelete, ExecuteDelete, StartRename, etc.
- search.rs: StartSearch, SearchNext, FuzzyFinder actions
- input.rs: ConfirmInput handling
- display.rs: Preview, Help, ToggleFocus, Refresh, etc.
- tests.rs: All existing tests preserved

main.rs (991 lines) -> src/app/ module + main.rs (81 lines):
- app/mod.rs: Module exports
- app/config.rs: CLI configuration and argument parsing
- app/preview.rs: PreviewState management
- app/event_loop.rs: Main event loop
- app/render.rs: Rendering helpers

All 286 tests pass. No functionality changes.
