+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "fileview"
-version = "1.4.0"
+version = "1.5.0"
 edition = "2021"
 description = "A minimal file tree UI for terminal emulators"
 authors = ["Hiro-Chiba"]
