+++ src/app/mod.rs
@@ -1,0 +1,13 @@
+//! Application module
+//!
+//! This module contains the main application logic, configuration,
+//! and event loop for FileView.
+
+mod config;
+mod event_loop;
+mod preview;
+mod render;
+
+pub use config::Config;
+pub use event_loop::{run_app, AppResult};
+pub use preview::PreviewState;
